requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.0.0
sendgrid==6.10.0
gunicorn==21.2.0
google-generativeai==0.3.2
